pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "philo-chat"
version = "0.3.0"
authors = [{ name = "Erfan Moosavi", email = "erfanmoosavi84@gmail.com" }]
description = "Chat with your favorite philosophers - Nietzsche, Socrates, and more!"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
  "fastapi[standard]==0.129.0",
  "openai>=1.0.0",
  "pydantic>=2.0.0",
  "python-dotenv>=1.0.1",
  "pyyaml>=6.0",
]

[project.scripts]
philo-chat = "src.cli.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"src" = ["data/*.json", "data/*.yaml", "py.typed"]
